implementacion inferza consola de empleado

+++ sql/crear-Empleado.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS empleados (
     empleado_id VARCHAR(20) NOT NULL,
     nombre VARCHAR(100) NOT NULL,
-    salario_base INTEGER NOT NULL
+    salario_base NUMERIC(10, 2) NOT NULL
 );
